Refactor: Extract ResetSection and update text per feedback
- Extracted ResetSection to app/client/connect/components/ResetSection.tsx.
- Updated button text to 'Reset Permissions & Settings' to align with actual UI intent and reviewer feedback.
- Updated all relevant unit and Playwright tests to match the new text.
- Ensured consistent border styling using theme.palette.divider.
